# backend/docs/swagger.yaml
definitions:
...
  model.Node:
    properties:
      _id:
        type: string
      create_ts:
        type: string
      description:
        type: string
      hostname:
        type: string
      ip:
        type: string
      is_master:
        description: 前端展示
        type: boolean
      key:
        description: 用于唯一标识节点，可能是mac地址，可能是ip地址
        type: string
      mac:
        type: string
      name:
        type: string
      port:
        type: string
      status:
        type: string
      update_ts:
        type: string
      update_ts_unix:
        type: integer
    type: object
...
  model.Schedule:
    properties:
      _id:
        type: string
      create_ts:
        type: string
      cron:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      entry_id:
        type: integer
      message:
        type: string
      name:
        type: string
      node_ids:
        items:
          type: string
        type: array
      nodes:
        items:
          $ref: '#/definitions/model.Node'
        type: array
      param:
        type: string
      run_type:
        type: string
      scrapy_log_level:
        type: string
      scrapy_spider:
        type: string
      spider_id:
        type: string
      spider_name:
        description: 前端展示
        type: string
      status:
        type: string
      update_ts:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /schedules:
    get:
      description: Get spider list
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page num
        in: query
        name: page_num
        type: string
      - description: page size
        in: query
        name: page_size
        type: string
      - description: keyword
        in: query
        name: keyword
        type: string
      - description: project_id
        in: query
        name: project_id
        type: string
      - description: type
        in: query
        name: type
        type: string
      - description: sort_key
        in: query
        name: sort_key
        type: string
      - description: sort_direction
        in: query
        name: sort_direction
        type: string
      - description: owner_type
        in: query
        name: owner_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get spider list
      tags:
      - spider
    put:
      consumes:
      - application/json
      description: Put schedule
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: schedule item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Schedule'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Put schedule
      tags:
      - schedule
  /schedules/{id}:
    delete:
      description: Delete schedule
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: schedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Delete schedule
      tags:
      - schedule
    get:
      description: Get schedule  by id
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: schedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get schedule by id
      tags:
      - schedule
    post:
      consumes:
      - application/json
      description: Post schedule
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: schedule id
        in: path
        name: id
        required: true
        type: string
      - description: schedule item
        in: body
        name: newItem
        required: true
        schema:
          $ref: '#/definitions/model.Schedule'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Post schedule
      tags:
      - schedule
  /schedules/{id}/disable:
    post:
      consumes:
      - application/json
      description: disable schedule
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: schedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: disable schedule
      tags:
      - schedule
  /schedules/{id}/enable:
    post:
      consumes:
      - application/json
      description: enable schedule
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: schedule id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: enable schedule
      tags:
      - schedule
...
  /stats/home:
    get:
      description: Get home stats
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get  home stats
      tags:
      - version
...
swagger: "2.0"